requirements.txt
google-generativeai
numpy
faiss-cpu
pandas
pyTelegramBotAPI